feat: Change Node-Traits to all accept mut-refs.
Considering a statefull node, non-mutable on_init, on_ready or
on_shutdown are not allowing any states to be stored or modified. Since
now one had to initialize this in the on_update method which does not
really make sense.

Breaking changes: This needs to be adapted for every implenetation of
on_init, on_ready and on_shutdown